# Improve performance of BigInteger.Multiply(large, small)