--- a/src/libraries/System.Runtime.Numerics/tests/BigInteger/multiply.cs
+++ b/src/libraries/System.Runtime.Numerics/tests/BigInteger/multiply.cs
@@ -205,6 +205,28 @@ namespace System.Numerics.Tests
         }
 
         [Fact]
+        public static void RunMultiplyKaratsubaBoundary()
+        {
+            Random random = new Random(s_seed);
+            byte[] tempByteArray1 = new byte[0];
+            byte[] tempByteArray2 = new byte[0];
+
+            // Multiply Method - One Large BigInteger
+            for (int i = 0; i < s_samples; i++)
+            {
+                for (int d1 = -2; d1 <= 2; d1++)
+                {
+                    tempByteArray1 = GetRandomByteArray(random, BigIntegerCalculator.MultiplyThreshold + d1);
+                    for (int d2 = -4; d2 <= 4; d2++)
+                    {
+                        tempByteArray2 = GetRandomByteArray(random, (BigIntegerCalculator.MultiplyThreshold + 1) * 2 + d2);
+                        VerifyMultiplyString(Print(tempByteArray1) + Print(tempByteArray2) + "bMultiply");
+                    }
+                }
+            }
+        }
+
+        [Fact]
         public static void RunMultiply_OnePositiveOneNegative()
         {
             Random random = new Random(s_seed);