BigInteger.Multiply is based on [Karatsuba algorithm](https://en.wikipedia.org/wiki/Karatsuba_algorithm). If implemented correctly, the computational complexity of multiply is $\Theta(n^{\log_2 3})$ where n is number of digits.

However, in the current implementation, it is not. This is because it the half of the smaller value is used when the larger one should be. 
https://github.com/dotnet/runtime/blob/353d5ead13cf66b81ad2579adb8f495afb0438a6/src/libraries/System.Runtime.Numerics/src/System/Numerics/BigIntegerCalculator.SquMul.cs#L214



## Benchmark
```

BenchmarkDotNet v0.13.8, Windows 11 (10.0.22621.2283/22H2/2022Update/SunValley2)
13th Gen Intel Core i5-13500, 1 CPU, 20 logical and 14 physical cores
.NET SDK 8.0.100-rc.1.23415.11
  [Host]   : .NET 8.0.0 (8.0.23.41404), X64 RyuJIT AVX2
  ShortRun : .NET 8.0.0 (8.0.23.41404), X64 RyuJIT AVX2

Job=ShortRun  Toolchain=.NET 8.0  IterationCount=3  
LaunchCount=1  WarmupCount=3  

```
| Method      | largeLength | smallLength | Mean          | Error       | StdDev    | Ratio | RatioSD |
|------------ |------------ |------------ |--------------:|------------:|----------:|------:|--------:|
| **PR_Multiply** | **1000**        | **1000**        |     **0.0375 ms** |   **0.0442 ms** | **0.0024 ms** |  **1.37** |    **0.09** |
| Multiply    | 1000        | 1000        |     0.0274 ms |   0.0002 ms | 0.0000 ms |  1.00 |    0.00 |
|             |             |             |               |             |           |       |         |
| **PR_Multiply** | **10000**       | **1000**        |     **0.3143 ms** |   **0.0750 ms** | **0.0041 ms** |  **0.58** |    **0.01** |
| Multiply    | 10000       | 1000        |     0.5409 ms |   0.0160 ms | 0.0009 ms |  1.00 |    0.00 |
|             |             |             |               |             |           |       |         |
| **PR_Multiply** | **10000**       | **10000**       |     **1.1306 ms** |   **0.2769 ms** | **0.0152 ms** |  **1.29** |    **0.02** |
| Multiply    | 10000       | 10000       |     0.8797 ms |   0.0351 ms | 0.0019 ms |  1.00 |    0.00 |
|             |             |             |               |             |           |       |         |
| **PR_Multiply** | **100000**      | **1000**        |     **3.0304 ms** |   **0.1180 ms** | **0.0065 ms** |  **0.52** |    **0.00** |
| Multiply    | 100000      | 1000        |     5.7730 ms |   0.6991 ms | 0.0383 ms |  1.00 |    0.00 |
|             |             |             |               |             |           |       |         |
| **PR_Multiply** | **100000**      | **10000**       |    **11.0477 ms** |   **0.7075 ms** | **0.0388 ms** |  **0.23** |    **0.00** |
| Multiply    | 100000      | 10000       |    48.9797 ms |   1.8604 ms | 0.1020 ms |  1.00 |    0.00 |
|             |             |             |               |             |           |       |         |
| **PR_Multiply** | **100000**      | **100000**      |    **39.4357 ms** |   **0.2381 ms** | **0.0131 ms** |  **1.25** |    **0.00** |
| Multiply    | 100000      | 100000      |    31.6607 ms |   0.2959 ms | 0.0162 ms |  1.00 |    0.00 |
|             |             |             |               |             |           |       |         |
| **PR_Multiply** | **1000000**     | **1000**        |    **30.2207 ms** |   **1.4899 ms** | **0.0817 ms** |  **0.48** |    **0.00** |
| Multiply    | 1000000     | 1000        |    63.0998 ms |  10.1509 ms | 0.5564 ms |  1.00 |    0.00 |
|             |             |             |               |             |           |       |         |
| **PR_Multiply** | **1000000**     | **10000**       |   **120.0384 ms** |   **8.3355 ms** | **0.4569 ms** |  **0.21** |    **0.00** |
| Multiply    | 1000000     | 10000       |   582.8313 ms |  39.3648 ms | 2.1577 ms |  1.00 |    0.00 |
|             |             |             |               |             |           |       |         |
| **PR_Multiply** | **1000000**     | **100000**      |   **448.9208 ms** |   **3.7954 ms** | **0.2080 ms** |  **0.09** |    **0.00** |
| Multiply    | 1000000     | 100000      | 5,091.4759 ms | 151.9431 ms | 8.3285 ms |  1.00 |    0.00 |
|             |             |             |               |             |           |       |         |
| **PR_Multiply** | **1000000**     | **995000**      | **1,608.9999 ms** |  **75.3231 ms** | **4.1287 ms** |  **1.09** |    **0.00** |
| Multiply    | 1000000     | 995000      | 1,482.7404 ms |  58.9864 ms | 3.2332 ms |  1.00 |    0.00 |
|             |             |             |               |             |           |       |         |
| **PR_Multiply** | **1000000**     | **1000000**     | **1,627.7224 ms** |  **63.8822 ms** | **3.5016 ms** |  **1.35** |    **0.01** |
| Multiply    | 1000000     | 1000000     | 1,209.0913 ms |  45.2731 ms | 2.4816 ms |  1.00 |    0.00 |

```cs
public class BigIntegerMultiplyBenchmark
{
    static ReadOnlySpan<byte> MakeBytes(int length)
    {
        var random = new Random(918);
        var bytes = new byte[length];
        random.NextBytes(bytes);
        return bytes;
    }

    public IEnumerable<object[]> LengthArguments()
    {
        var lengths = new int[] { 1000, 10000, 100000, 1000000 };
        for (int i = lengths.Length - 1; i >= 0; i--)
        {
            if (i == lengths.Length - 1)
            {
                yield return new object[] { lengths[i], (int)(0.995 * lengths[i]), };
            }
            for (int j = i; j >= 0; j--)
            {
                yield return new object[] { lengths[i], lengths[j], };
            }
        }
    }

    [Benchmark]
    [ArgumentsSource(nameof(LengthArguments))]
    public PrBigInteger PR_Multiply(int largeLength, int smallLength)
    {
        return (new PrBigInteger(MakeBytes(smallLength)) * new PrBigInteger(MakeBytes(largeLength)));
    }

    [Benchmark(Baseline = true)]
    [ArgumentsSource(nameof(LengthArguments))]
    public BigInteger Multiply(int largeLength, int smallLength)
    {
        return (new BigInteger(MakeBytes(smallLength)) * new BigInteger(MakeBytes(largeLength)));
    }
}
```