@tannergooding 2023-09-20
Could you add the benchmark to https://github.com/dotnet/performance/blob/main/src/benchmarks/micro/libraries/System.Runtime.Numerics/Perf.BigInteger.cs (or ensure the existing `Multiply` benchmark sufficiently covers the scenario)?

Changes in general LGTM, just want to ensure we have some perf numbers before we merge so it can be correctly tracked in our historical data.